pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "codegraph"
version = "0.1.0"
description = "Give AI agents a graph view of your codebase. Stop wasting context on exploration."
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Agam Gupta" }]
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "networkx>=3.2",
    "strawberry-graphql[fastapi]>=0.235.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic-settings>=2.2.0",
    "structlog>=24.1.0",
    "mcp>=1.0.0",
    "sentence-transformers>=2.7.0",
    "faiss-cpu>=1.8.0",
    "tree-sitter>=0.22.0",
    "tree-sitter-python>=0.22.0",
    "tree-sitter-javascript>=0.22.0",
    "tree-sitter-typescript>=0.22.0",
    "aiosqlite>=0.20.0",
    "click>=8.1.0",
    "pathspec>=0.12.0",
    "rich>=13.7.0",
    "httpx>=0.27.0",
]

[project.scripts]
codegraph = "codegraph.main:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.10.0",
    "ruff>=0.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/codegraph"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=codegraph --cov-report=term-missing"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py312"
